+++ backend/src/main/java/application/service/TaskService.java
@@ -20,14 +20,14 @@ public class TaskService {
         taskRepository.save(task);
     }
 
-    public Task setPreviousTask(Integer id, Task previousTask) {
+    public Task setPreviousTask(Long id, Task previousTask) {
         Task task = taskRepository.findById(id).get();
         task.setPreviousTask(previousTask);
 
         return taskRepository.save(task);
     }
 
-    public Task updateTask(Integer id, Task newTask) {
+    public Task updateTask(Long id, Task newTask) {
         Task task = taskRepository.findById(id).get();
         task.setName(newTask.getName());
         task.setDescription(newTask.getDescription());
@@ -42,7 +42,7 @@ public class TaskService {
         return taskRepository.save(task);
     }
 
-    public void deleteTask(Integer id) {
+    public void deleteTask(Long id) {
         taskRepository.deleteById(id);
     }
 }
